+++ src/labbench/paramattr/__init__.py
@@ -10,7 +10,6 @@ from ._bases import (
     observe,
     unobserve,
 )
-
 from ._key_adapters import message_keying, visa_keying
 
 for _obj in dict(locals()).values():