+++ src/labbench/paramattr/method.py
@@ -1,6 +1,6 @@
 
+
 from . import _bases, _types
-import typing as _typing
 from ._bases import Method